add Artifact class

+++ src/core/impl/writable/writable_with_offset.cpp
@@ -2,7 +2,7 @@
 
 namespace ark {
 
-WritableWithOffset::WritableWithOffset(Writable& writable, const size_t offset)
+WritableWithOffset::WritableWithOffset(Writable& writable, const uint32_t offset)
     : _writable(writable), _offset(offset)
 {
 }
